+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 * Added `integration [start]`
 * Added `deploy` via larga
+* Hard reset local feature and bugfix branches to remote versions before adding
 
 ## 0.1.0
 * Initial release